Reflection-config plugin: Ability to deal with relocated classes
When a project's classes are (later) relocated, for example via the
shadow plugin, the generated `reflection-config.json` should also
reference the relocated class names.

It is quite tricky, and actually subject to the Shadow plugin's
internals, to inquire the relocated class name for a "source" class
name. It would also make the reflection-config plugin depend on the
shadow plugin and potentially cause plugin class path issues in Gradle.

The solution here is to provide a map of regex replacements.